Start themeing site and format everything
Started using Tailwind to theme the site up. It's a bit of a learning
curve, but I'm hoping to get a better result than if I had done it by
hand. Tailwind is nice so far.

Also figured out how to get Prettier to work on EEx files so got that
all setup and formatted everything. That's going to make my life much
nicer.