Progress on voice actors
Changed some things around in the database to follow best practices, and
update a number of minor things to catch some bugs and error louder.

Limited the voice actors to the first 5 characters, and the first VA for
each to greatly cut down on the number of entries and overall simplify
things. This solves my biggest problem with storing/listing VAs.