Switch Etag cache to using Postgres
The etag cache now saves to the Postgres database instead of using ETS.
This is probably what I should have done from the start but I wanted to
use ETS for something.

This way the cache is persistent and stored for all replicas (if there
are any) which should help decrease the load on Jikan and speed up the
site.

Also this means that seeding the initial database is easier and can
include cached pages.