Seed data and fix issues
Riding on the tails of the postgres caching switch is some seed data to
get the database started with. The seed scripts are a little iffy due to
Jikan rate limits, but they should be alright.

With that I discovered a few places where the code is brittle, and some
more improvements that could be made to the ETag cache and Jikan modules
to fix bugs and cleanup code.