Initial REST API and Sitemap
Created the initial REST API which follows the documented
specifications. This unearthed a few bugs in the database code so
cleaned those up and added the new stuff I needed.

Had to update the router too and fixed a bad issue there too

Also added a basic generated sitemap which will be useful for SEO later

Progress towards #27
Progress towards #29