Finished splitting characters and VAs
Characters and Voice Actors are now fully their own tables.
The Jikan module had to be almost totally re-written to adjust for this,
but it now flows much better and pulls all the information better.

Everything is pulled and coalesced into known structures, and then
inserted into the database in one transaction.

Progress towards #51